git-svn-id: https://svn.osdn.net/svnroot/ttssh2/branches/cmake@7392 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/common/tttypes.h
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 1994-1998 T. Teranishi
- * (C) 2004-2018 TeraTerm Project
+ * (C) 2004-2019 TeraTerm Project
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -550,8 +550,8 @@ struct tttset {
 	/* protocol used in connect() */
 	int ProtocolFamily;
 	char MouseCursorName[16];
-	int AlphaBlendActive;
-	int AlphaBlendInactive;
+	BYTE AlphaBlendActive;
+	BYTE AlphaBlendInactive;
 	char CygwinDirectory[MAX_PATH];
 #define DEFAULT_LOCALE "japanese"
 	char Locale[80];
